Prevented preseed from creating a dtg account. By removing this account,we can also remove the hashed password from the preseed, and so publish on github. Therefore, we can also point new VM installs at a github copy of the preseed config, and remove apache from dhcp.
It turns out that Debian's preseed blah doesn't expect you to want to install Debian/Ubuntu with root logins disabled, and no user account. Usually this is pretty pointless, asyou cannot login. However we have magic that creates user accounts before the login screen appears. To work around this we set the password entry for root in /etc/shadow  to !!, which is password expired. According to the Interwebz, */LK won't work. The same source states that root cannot login with an expired password.