Make root's password *, not !!.
This patch changes the entry in /etc/shadow for root, so that it's
password is *, not !!. The advantage of this change is that we can
then SSH in as root. More specifically root@husky0 can. This is done by
means of the key inserted into ~root/.ssh/authorized_keys by the VM boot
script, if the hostname is puppy*.

When the machine is rebooted, if the hostname is no longer puppy*,
the key is then deleted. At this point, it ought to no longer be
possible to login as root:
0) The * entry made by this patch means that the user has no valid
password that can be used to login
1) Root has no keys in its authorized_keys, unless someone has
pratted about with it. Then all bets are off, and you are being stupid.