Change from expert recipe to atomic.
We want (and currently have) vms with one disk containing one parition
containing the OS. We then have all data on a second disk.

Up until now we have been trying to write a recipe that describes `one
partition with everything', but done it a bit badly. Partman already
has this functionality built in, and requires 1 LoC, not the 20 odd we had.